Better_Debug/requirements.txt
pyserial>=3.5
pytest>=8.0
